Starts process of generalising the plugin framework
* Renames komodo-plugin-teiid-framework to komodo-plugin-framework
  so other plugins can make use of it.
