+++ komodo-parent/pom.xml
@@ -370,7 +370,7 @@
 			</dependency>
 			<dependency>
 				<groupId>org.komodo.plugins</groupId>
-				<artifactId>komodo-plugin-teiid-framework</artifactId>
+				<artifactId>komodo-plugin-framework</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 			<dependency>
