+++ plugins/komodo-plugin-framework/.gitignore
