+++ plugins/komodo-plugin-framework/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>komodo-plugin-teiid-framework</name>
+	<name>komodo-plugin-framework</name>
 	<comment></comment>
 	<projects>
 	</projects>
