+++ plugins/komodo-plugin-framework/src/main/java/org/komodo/plugin/framework/teiid/AbstractTeiidInstance.java
@@ -19,7 +19,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
  * 02110-1301 USA.
  */
-package org.komodo.teiid.framework;
+package org.komodo.plugin.framework.teiid;
 
 import java.io.File;
 import java.io.FileInputStream;
