+++ plugins/komodo-plugin-framework/src/test/java/org/komodo/teiid/AbstractTestTeiidInstance.java
