+++ plugins/komodo-plugin-service/pom.xml
@@ -13,6 +13,9 @@
 	<properties>
 		<version.apache.felix.framework>5.4.0</version.apache.felix.framework>
 		<version.apache.felix.fileinstall>3.5.2</version.apache.felix.fileinstall>
+		<includePluginGroupIds>
+			org.komodo.plugins.teiid,
+		</includePluginGroupIds>
 	</properties>
 
 	<dependencyManagement>
@@ -41,7 +44,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.komodo.plugins</groupId>
-			<artifactId>komodo-plugin-teiid-framework</artifactId>
+			<artifactId>komodo-plugin-framework</artifactId>
 		</dependency>
 
 		<dependency>
@@ -163,7 +166,7 @@
 							<goal>copy-dependencies</goal>
 						</goals>
 						<configuration>
-							<includeGroupIds>org.komodo.plugins.teiid</includeGroupIds>
+							<includeGroupIds>${includePluginGroupIds}</includeGroupIds>
 							<outputDirectory>${project.basedir}/src/main/resources/bundles</outputDirectory>
 							<stripVersion>true</stripVersion>
 							<overWriteReleases>false</overWriteReleases>
