+++ plugins/teiid-8.11/pom.xml
@@ -108,7 +108,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.komodo.plugins</groupId>
-			<artifactId>komodo-plugin-teiid-framework</artifactId>
+			<artifactId>komodo-plugin-framework</artifactId>
 			<scope>provided</scope>
 		</dependency>
 		<dependency>
@@ -173,7 +173,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.komodo.plugins</groupId>
-			<artifactId>komodo-plugin-teiid-framework</artifactId>
+			<artifactId>komodo-plugin-framework</artifactId>
 			<type>test-jar</type>
 			<scope>test</scope>
 		</dependency>
