+++ plugins/teiid-8.11/src/main/java/org/komodo/teiid/TeiidVdbImpl.java
@@ -27,8 +27,8 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Properties;
+import org.komodo.plugin.framework.teiid.Messages;
 import org.komodo.spi.runtime.TeiidVdb;
-import org.komodo.teiid.framework.Messages;
 import org.teiid.adminapi.Model;
 import org.teiid.adminapi.VDB;
 import org.teiid.adminapi.VDB.Status;
