+++ plugins/teiid-8.12/pom.xml
@@ -107,7 +107,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.komodo.plugins</groupId>
-			<artifactId>komodo-plugin-teiid-framework</artifactId>
+			<artifactId>komodo-plugin-framework</artifactId>
 			<scope>provided</scope>
 		</dependency>
 		<dependency>
@@ -172,7 +172,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.komodo.plugins</groupId>
-			<artifactId>komodo-plugin-teiid-framework</artifactId>
+			<artifactId>komodo-plugin-framework</artifactId>
 			<type>test-jar</type>
 			<scope>test</scope>
 		</dependency>
