+++ plugins/teiid-8.12/src/main/java/org/komodo/teiid/DataTypeManagerImpl.java
@@ -22,8 +22,8 @@
 package org.komodo.teiid;
 
 import java.util.Set;
+import org.komodo.plugin.framework.teiid.AbstractDataTypeManager;
 import org.komodo.spi.runtime.version.TeiidVersion;
-import org.komodo.teiid.framework.AbstractDataTypeManager;
 import org.teiid.core.types.DataTypeManager;
 
 public class DataTypeManagerImpl extends AbstractDataTypeManager {
