+++ plugins/teiid-8.12/src/main/java/org/komodo/teiid/JbossExtensions.java
@@ -32,9 +32,9 @@ import org.jboss.as.cli.Util;
 import org.jboss.as.cli.operation.impl.DefaultOperationRequestBuilder;
 import org.jboss.as.controller.client.ModelControllerClient;
 import org.jboss.dmr.ModelNode;
+import org.komodo.plugin.framework.teiid.ExtensionConstants;
+import org.komodo.plugin.framework.teiid.Messages;
 import org.komodo.spi.runtime.DataSourceDriver;
-import org.komodo.teiid.framework.ExtensionConstants;
-import org.komodo.teiid.framework.Messages;
 import org.teiid.adminapi.Admin;
 import org.teiid.adminapi.jboss.MetadataMapper;
