+++ plugins/teiid-8.12/src/main/java/org/komodo/teiid/NodeGenerator.java
@@ -24,6 +24,7 @@ package org.komodo.teiid;
 import java.util.List;
 import java.util.Map;
 import javax.jcr.Node;
+import org.komodo.plugin.framework.teiid.AbstractNodeGenerator;
 import org.komodo.spi.constants.StringConstants;
 import org.komodo.spi.lexicon.TeiidSqlLexicon;
 import org.komodo.spi.query.CriteriaOperator;
@@ -32,7 +33,6 @@ import org.komodo.spi.query.JoinTypeTypes;
 import org.komodo.spi.runtime.version.TeiidVersion;
 import org.komodo.spi.type.DataTypeManager;
 import org.komodo.spi.type.DataTypeManager.DataTypeName;
-import org.komodo.teiid.framework.AbstractNodeGenerator;
 import org.teiid.language.Like.MatchMode;
 import org.teiid.language.SortSpecification.NullOrdering;
 import org.teiid.query.sql.LanguageObject;
