+++ plugins/teiid-8.12/src/main/java/org/komodo/teiid/TeiidTranslatorImpl.java
@@ -24,8 +24,8 @@ package org.komodo.teiid;
 import java.util.Map.Entry;
 import java.util.Properties;
 import java.util.Set;
+import org.komodo.plugin.framework.teiid.Messages;
 import org.komodo.spi.runtime.TeiidTranslator;
-import org.komodo.teiid.framework.Messages;
 import org.teiid.adminapi.AdminObject;
 import org.teiid.adminapi.Translator;
 import org.teiid.core.util.ArgCheck;
