+++ plugins/teiid-template/template/src/main/java/org/komodo/teiid/DataTypeManagerImpl.java
@@ -24,7 +24,7 @@ package org.komodo.teiid;
 import java.util.Set;
 import org.komodo.spi.runtime.version.TeiidVersion;
 import org.komodo.spi.type.DataTypeManager.DataTypeName;
-import org.komodo.teiid.framework.AbstractDataTypeManager;
+import org.komodo.plugin.framework.teiid.AbstractDataTypeManager;
 import org.teiid.core.types.DataTypeManager;
 
 public class DataTypeManagerImpl extends AbstractDataTypeManager {
