+++ plugins/teiid-template/template/src/main/java/org/komodo/teiid/TeiidInstanceImpl.java
@@ -44,8 +44,8 @@ import org.komodo.spi.runtime.TeiidPropertyDefinition;
 import org.komodo.spi.runtime.TeiidTranslator;
 import org.komodo.spi.runtime.TeiidVdb;
 import org.komodo.spi.runtime.version.TeiidVersion;
-import org.komodo.teiid.framework.AbstractTeiidInstance;
-import org.komodo.teiid.framework.Messages;
+import org.komodo.plugin.framework.teiid.AbstractTeiidInstance;
+import org.komodo.plugin.framework.teiid.Messages;
 import org.teiid.adminapi.Admin;
 import org.teiid.adminapi.PropertyDefinition;
 import org.teiid.adminapi.Translator;
