+++ plugins/teiid-template/template/src/main/java/org/komodo/teiid/TeiidServiceImpl.java
@@ -27,8 +27,8 @@ import org.komodo.spi.runtime.TeiidJdbcInfo;
 import org.komodo.spi.runtime.TeiidParent;
 import org.komodo.spi.runtime.version.DefaultTeiidVersion;
 import org.komodo.spi.type.DataTypeManager;
-import org.komodo.teiid.framework.AbstractTeiidService;
-import org.komodo.teiid.framework.Messages;
+import org.komodo.plugin.framework.teiid.AbstractTeiidService;
+import org.komodo.plugin.framework.teiid.Messages;
 import org.teiid.core.util.ApplicationInfo;
 import org.teiid.query.parser.QueryParser;
 import org.teiid.query.sql.lang.Command;
